+++ src/roq/samples/example-6/application.h
@@ -1,0 +1,27 @@
+/* Copyright (c) 2017-2021, Hans Erik Thrane */
+
+#pragma once
+
+#include <string_view>
+
+#include "roq/service.h"
+
+#include "roq/span.h"
+
+namespace roq {
+namespace samples {
+namespace example_6 {
+
+class Application final : public Service {
+ public:
+  using Service::Service;
+
+ protected:
+  int main_helper(const roq::span<std::string_view> &args);
+
+  int main(int argc, char **argv) override;
+};
+
+}  // namespace example_6
+}  // namespace samples
+}  // namespace roq