+++ src/roq/samples/example-6/flags.h
@@ -1,0 +1,15 @@
+/* Copyright (c) 2017-2021, Hans Erik Thrane */
+
+#pragma once
+
+#include "roq/samples/example-6/flags/flags.h"
+
+namespace roq {
+namespace samples {
+namespace example_6 {
+
+using flags::Flags;
+
+}  // namespace example_6
+}  // namespace samples
+}  // namespace roq